adding Status Codes

+++ src/main/java/com/chauhan/springbootemployeewebproject/springbootemployeewebproject/services/EmployeeService.java
@@ -10,6 +10,7 @@ import org.springframework.data.util.ReflectionUtils;
 import java.lang.reflect.Field;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 @Service
 public class EmployeeService {
@@ -21,9 +22,13 @@ public class EmployeeService {
         this.modelMapper = modelMapper;
     }
 
-    public EmployeeDTO getEmployeeById(Long id){
-        EmployeeEntity employeeEntity = employeeRepository.findById(id).orElse(null);
-        return modelMapper.map(employeeEntity,EmployeeDTO.class);
+    public Optional<EmployeeDTO> getEmployeeById(Long id){
+//        EmployeeEntity employeeEntity = employeeRepository.findById(id).orElse(null);
+//        return modelMapper.map(employeeEntity,EmployeeDTO.class);
+
+        return employeeRepository
+                .findById(id)
+                .map(employeeEntity -> modelMapper.map(employeeEntity,EmployeeDTO.class));
     }
 
     public List<EmployeeDTO>getAllEmployees(){
@@ -40,14 +45,35 @@ public class EmployeeService {
     }
 
     public EmployeeDTO updateEmployeeById(Long id, EmployeeDTO employeeDTO){
-        EmployeeEntity employeeEntity = modelMapper.map(employeeDTO,EmployeeEntity.class);
-        employeeEntity.setId(id);
-        EmployeeEntity savedEntity = employeeRepository.save(employeeEntity);
-        return modelMapper.map(savedEntity,EmployeeDTO.class);
+        boolean exists = isExistsByEmployeeId(id);
+        if(!exists) return null;
+
+        EmployeeEntity employeeEntity = employeeRepository.findById(id).get();
+
+        // Update only the necessary fields
+        // Update only non-null fields from employeeDTO to employeeEntity
+        if (employeeDTO.getName() != null) {
+            employeeEntity.setName(employeeDTO.getName());
+        }
+        if (employeeDTO.getAge() != null) {
+            employeeEntity.setAge(employeeDTO.getAge());
+        }
+        if (employeeDTO.getDateOfJoining() != null) {
+            employeeEntity.setDateOfJoining(employeeDTO.getDateOfJoining());
+        }
+        if (employeeDTO.getIsActive() != null) {
+            employeeEntity.setIsActive(employeeDTO.getIsActive());
+        }
+
+        EmployeeEntity updatedEmployee = employeeRepository.save(employeeEntity);
+        return modelMapper.map(updatedEmployee, EmployeeDTO.class);
     }
 
-    public void deleteEmployeeById(Long id){
+    public boolean deleteEmployeeById(Long id){
+        boolean exists = isExistsByEmployeeId(id);
+        if(!exists) return false;
         employeeRepository.deleteById(id);
+        return true;
     }
 
     public boolean isExistsByEmployeeId(Long id){
